Fixes 6
